DTSPB-2794 fixing concurrency issue
