+++ src/main/java/uk/gov/hmcts/reform/migration/CaseMigrationProcessor.java
@@ -167,52 +167,6 @@ public class CaseMigrationProcessor {
         }
     }
 
-    private boolean setCaseToHandedOffToLegacySite(Map<String, Object> caseData) {
-        if (caseData.containsKey("applicationType") && caseData.get("applicationType").equals("Solicitor")
-            && (caseData.containsKey("titleAndClearingType")
-                && (caseData.get("titleAndClearingType").equals("TCTTrustCorpResWithSDJ")
-                    || caseData.get("titleAndClearingType").equals("TCTTrustCorpResWithApp")))
-        ) {
-            return true;
-        }
-        if (caseData.containsKey("applicationType") && caseData.get("applicationType").equals("Solicitor")
-            && (caseData.containsKey("caseType")
-                && (caseData.get("caseType").equals("gop")
-                || caseData.get("caseType").equals("admonWill")
-                || caseData.get("caseType").equals("intestacy")))
-            && (caseData.containsKey("deceasedDomicileInEngWales")
-                && caseData.get("deceasedDomicileInEngWales").equals("No"))
-        ) {
-            return true;
-        }
-        if (caseData.containsKey("applicationType") && caseData.get("applicationType").equals("Solicitor")
-            && (caseData.containsKey("caseType")
-                && (caseData.get("caseType").equals("gop")
-                || caseData.get("caseType").equals("admonWill")
-                || caseData.get("caseType").equals("intestacy")))
-            && (caseData.containsKey("willAccessOriginal") && caseData.get("willAccessOriginal").equals("No"))
-            && (caseData.containsKey("willAccessNotarial") && caseData.get("willAccessNotarial").equals("Yes"))
-        ) {
-            return true;
-        }
-        if (caseData.containsKey("applicationType") && caseData.get("applicationType").equals("Solicitor")
-            && caseData.containsKey("caseType") && caseData.get("caseType").equals("intestacy")
-            && caseData.containsKey("solsApplicantRelationshipToDeceased")
-            && caseData.get("solsApplicantRelationshipToDeceased").equals("Yes")
-        ) {
-            return true;
-        }
-        if (caseData.containsKey("applicationType") && caseData.get("applicationType").equals("Personal")
-            && caseData.containsKey("caseType") && caseData.get("caseType").equals("intestacy")
-            && caseData.containsKey("primaryApplicantRelationshipToDeceased")
-            && caseData.get("primaryApplicantRelationshipToDeceased").equals("adoptedChild")
-            && caseData.containsKey("primaryApplicantAdoptionInEnglandOrWales")
-            && caseData.get("primaryApplicantAdoptionInEnglandOrWales").equals("Yes")
-        ) {
-            return true;
-        }
-        return false;
-    }
 
     private void updateCase(String authorisation, String caseType, CaseDetails caseDetails) {
         if (dataMigrationService.accepts().test(caseDetails)) {
@@ -221,11 +175,6 @@ public class CaseMigrationProcessor {
             try {
                 log.debug("Case data before: {}", caseDetails.getData());
                 Map<String, Object> caseData = caseDetails.getData();
-                if (setCaseToHandedOffToLegacySite(caseData)) {
-                    caseData.put("caseHandedOffToLegacySite","Yes");
-                } else {
-                    caseData.put("caseHandedOffToLegacySite","No");
-                }
                 coreCaseDataService.update(
                     authorisation,
                     EVENT_ID,
